modify: 구조 변경 - '요약'용 API 테스트 전 정리
(기본API 테스트 완 - 통신 성공)
- 요약용 기능을, 기존 핸들러 모듈과 독립적인 핸들러 모듈로 분리.
- DSPy의 기본 설정부를, 핸들러를 담는 패키지의 `__init__`으로 이동.
- 안건 비교 소스코드를 위한 준비.
- 프로젝트 가독성과 이해도 향상을 위한 구조 및 이름 변경.

+ 추가) RunPod 서버리스 API를 위한 Dockerfile 준비, 하지만 미완 상태.

이후, 현재 바뀐 구조가 도커 이미지에서도 정상 작동할 수 있도록
`Dockerfile_custom_api`를 손봐야 한다.
