+++ Dockerfile_runpod_api
@@ -1,0 +1,17 @@
+# Include Python
+FROM python:3.12
+
+# Set the current working directory to `/code`
+WORKDIR /code
+
+# Copy the file with the requirements to the `/code` directory
+COPY ./requirements.txt /code/requirements.txt
+
+# Install the package dependencies in the requirements file
+RUN pip install --no-cache-dir --upgrade -r /code/requirements.txt
+
+# Add your file
+ADD whatever.py .
+
+# Call your file when your container starts
+CMD [ "python", "-u", "/whatever.py" ]
