+++ src/api_custom/routers/_test_lm_comm.py
@@ -1,10 +1,8 @@
 from fastapi import APIRouter
-from api_custom.nlp_link.model_handler import BasicHandler
-# from app.config import MODEL_PATH
+from api_custom.nlp_link.model_handlers._basic_handler import BasicHandler
 
 router = APIRouter()
 
-# model_handler = ModelHandler(MODEL_PATH)
 model_handler = BasicHandler()
 
 @router.post("/lm_greet")
@@ -16,7 +14,3 @@ async def greet():
 async def answer(question: dict):
     answering = model_handler.custom_question_answer(question)
     return {"응답": answering}
-
-@router.post("/summarize")
-async def summarize(transcript):
-    summary = model_handler
