+++ src/nlp_core/README.md
@@ -1,0 +1,4 @@
+
+- `metric/`:
+- `model_handlers/`: 
+- `__init__.py`: 언어 모델 정보와 같은 기본적인 설정을 수행한다.
